move bluesky_kafka.kafka to bluesky_kafka
